eliminate unnecessary boxing and unboxing

+++ src/main/java/de/jarnbjo/util/io/ByteArrayBitInputStream.java
@@ -137,7 +137,7 @@ public class ByteArrayBitInputStream implements BitInputStream {
          }
          root=(currentByte&(1<<(bitIndex++)))!=0?root.o1:root.o0;
       }
-      return root.value.intValue();
+      return root.value;
    }
 
    public long getLong(int bits) throws IOException {
